+++ coretrace/simulation_runner/optix_runner/OptixCSP/src/core/soltrace_system.h
@@ -142,7 +142,7 @@ namespace OptixCSP
         size_t get_num_groups() const { return my_groups.size(); }
         void set_groups(const std::vector<int32_t>& groups) { m_groups = groups; }
         size_t get_num_groups() const { return m_groups.size(); }
-        int8_t get_group(int32_t element_id);
+        int32_t get_group(int32_t element_id);
 
     private:
         // m_verbose and m_state must be declared before the shared_ptr managers so
